pyproject.toml
[project]
name = "my-wagtail-fastapi-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "fastapi>=0.115.12",
    "pip-audit>=2.9.0",
    "python-dotenv>=1.1.0",
    "stripe>=12.2.0",
    "uvicorn>=0.34.3",
    "wagtail>=7.0.1",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
    "black>=23.7.0",
    "pre-commit>=3.6.0",
]
test = [
    "pytest>=8.4.0",
    "pytest-django>=4.8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.26.0",
    "pytest-cov>=4.0.0",
    "factory-boy>=3.3.0",
    "responses>=0.24.0",
    "freezegun>=1.5.0",
]
e2e = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.3",
    "httpx>=0.26.0",
    "pytest-base-url>=2.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "requests>=2.31.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_project.totonoe_template.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=fastapi_app",
    "--cov=blog",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["fastapi_app", "blog"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "*/staticfiles/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# ========================================
# Code Quality Tools Configuration
# ========================================

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "DJ", "SIM", "RUF"]
ignore = [
    "E501",  # line too long (handled by black)
    "E203",  # whitespace before ':' (handled by black)
    "E402",  # module level import not at top of file (Django)
    "F403",  # star imports (used in Django settings)
    "F405",  # may be undefined from star imports (Django settings)
    "F401",  # imported but unused (Django settings)
    "D100",  # missing docstring in public module
    "D101",  # missing docstring in public class
    "D102",  # missing docstring in public method
    "D103",  # missing docstring in public function
    "D104",  # missing docstring in public package
    "RUF002", # docstring contains ambiguous characters
    "RUF003", # comment contains ambiguous characters
    "B904",   # raise exceptions with from
    "SIM102", # nested if statements
    "SIM117", # multiple with statements
    "UP035",  # deprecated typing imports
    "RUF013", # implicit Optional
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "migrations",
    "node_modules",
    "staticfiles",
    "tests/e2e/",  # E2E tests excluded from strict formatting
]

[tool.ruff.per-file-ignores]
"tests/*" = ["D"]
"*/migrations/*" = ["ALL"]
"main_asgi.py" = ["E402"]  # Django setup needs to be before imports
"tests/conftest.py" = ["E402"]  # Django setup needs to be before imports

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E501", "E402", "F403", "F405", "F401", "E231", "E272", "E702", "E202", "E201", "E221"]
per-file-ignores = [
    "*/settings/*.py:F403,F405,F401",
    "main_asgi.py:E402",
    "tests/conftest.py:E402",
    "tests/e2e/*:ALL",
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "migrations",
    "staticfiles",
    "tests/e2e/",
]
